python-project/requirements.txt
pyspark==3.5.0
pandas==2.0.2
pyarrow==14.0.1
delta-spark==3.1.0
pytest==8.1.1
assertpy==1.1
pytest-clarity==1.0.1
great-expectations==0.18.12